+++ lib/src/io_client.dart
@@ -8,6 +8,7 @@ import 'base_client.dart';
 import 'base_request.dart';
 import 'exception.dart';
 import 'io_streamed_response.dart';
+import 'progress.dart';
 
 /// Create an [IOClient].
 ///
@@ -23,7 +24,10 @@ class IOClient extends BaseClient {
 
   /// Sends an HTTP request and asynchronously returns the response.
   @override
-  Future<IOStreamedResponse> send(BaseRequest request) async {
+  Future<IOStreamedResponse> send(
+    BaseRequest request, {
+    Progress? onSendProgress,
+  }) async {
     var stream = request.finalize();
 
     try {
@@ -36,7 +40,24 @@ class IOClient extends BaseClient {
         ioRequest.headers.set(name, value);
       });
 
-      var response = await stream.pipe(ioRequest) as HttpClientResponse;
+      HttpClientResponse response;
+      if (onSendProgress != null) {
+        var loaded = 0;
+        onSendProgress(loaded, request.contentLength);
+
+        await ioRequest.addStream(
+          stream.map(
+            (chunk) {
+              loaded += chunk.length;
+              onSendProgress(loaded, request.contentLength);
+              return chunk;
+            },
+          ),
+        );
+        response = await ioRequest.close();
+      } else {
+        response = await stream.pipe(ioRequest) as HttpClientResponse;
+      }
 
       var headers = <String, String>{};
       response.headers.forEach((key, values) {