+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.13.4-dev
 
+* Add `onSendProgress` callback
+
 ## 0.13.3
 
 * Validate that the `method` parameter of BaseRequest is a valid "token".